### 2023-04-05 #5342 @editorialbot
[PRE REVIEW]: GMP-Featurizer: A parallelized Python package for efficiently computing the Gaussian Multipole features of atomic systems
**Submitting author:** @RayLei-TRI (<a href="http://orcid.org/0000-0002-2333-9205">Xiangyun Lei</a>)
**Repository:** https://github.com/TRI-AMDD/GMP-featurizer/
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v1.0.0
