### 2023-04-05 #5342 @Kevin-Mattheus-Moerman
@RayLei-TRI after inspection of your repository it looks to me like you license is currently not correctly implemented. YOu should follow the instructions provided here:  https://www.apache.org/licenses/LICENSE-2.0, i.e. Your `LICENSE` file should include the long text given there, and then you can include the "_boilerplate_" text "To apply the Apache License to specific files in your work". Once you make this change, GitHub will also properly detect your license as being `Apache 2.0`.
