### 2023-04-05 #5342 @Kevin-Mattheus-Moerman
@openjournals/dev how come our system didn't produce a warning on the license here?
