+++ paper.bib
@@ -8,6 +8,7 @@
     number = {12},
     year = {2017},
     title = {Machine learning unifies the modeling of materials and molecules},
+    doi = {10.1126/sciadv.1701816},
     copyright = {http://arxiv.org/licenses/nonexclusive-distrib/1.0},
     language = {eng},
     address = {United States},
@@ -66,6 +67,7 @@
     number = {14},
     year = {2018},
     title = {Crystal Graph Convolutional Neural Networks for an Accurate and Interpretable Prediction of Material Properties},
+    doi = {10.1103/PhysRevLett.120.145301},
     copyright = {http://arxiv.org/licenses/nonexclusive-distrib/1.0},
     language = {eng},
     address = {United States},
@@ -81,7 +83,7 @@
     pages = {314-319},
     year = {2013},
     issn = {0927-0256},
-    doi = {https://doi.org/10.1016/j.commatsci.2012.10.028},
+    doi = {10.1016/j.commatsci.2012.10.028},
     url = {https://www.sciencedirect.com/science/article/pii/S0927025612006295},
     author = {Shyue Ping Ong and William Davidson Richards and Anubhav Jain and Geoffroy Hautier and Michael Kocher and Shreyas Cholia and Dan Gunter and Vincent L. Chevrier and Kristin A. Persson and Gerbrand Ceder},
     keywords = {Materials, Project, Design, Thermodynamics, High-throughput},
@@ -96,6 +98,7 @@
     Peterson and Carsten Rostgaard and Jakob Schiøtz and Ole Schütt and Mikkel Strange and Kristian S Thygesen and Tejs
     Vegge and Lasse Vilhelmsen and Michael Walter and Zhenhua Zeng and Karsten W Jacobsen},
     title={The atomic simulation environment—a Python library for working with atoms},
+    doi = {10.1088/1361-648X/aa680e},
     journal={Journal of Physics: Condensed Matter},
     volume={29},
     number={27},
