### 2023-05-01 #5342 @Kevin-Mattheus-Moerman
@richardjgowers, the person @isdanni is not an editor for JOSS, I believe they are offering to review, is that right @isdanni?
