### 2023-05-10 f838cd0
add scaling constant bsaed on both probes
