+++ tests/featurizer_test.py
@@ -26,30 +26,21 @@ class FeatruizerTest(unittest.TestCase):
         )
 
         featurizer_fp_deriv = GMPFeaturizer(
-            GMPs=GMPs,
-            calc_derivatives=True,
-            calc_occ_derivatives=False,
-            verbose=False,
+            GMPs=GMPs, calc_derivatives=True, calc_occ_derivatives=False, verbose=False,
         )
         features_fp_deriv = featurizer_fp_deriv.prepare_features(
             images, cores=cores, converter=converter
         )
 
         featurizer_occ_deriv = GMPFeaturizer(
-            GMPs=GMPs,
-            calc_derivatives=False,
-            calc_occ_derivatives=True,
-            verbose=False,
+            GMPs=GMPs, calc_derivatives=False, calc_occ_derivatives=True, verbose=False,
         )
         features_occ_deriv = featurizer_occ_deriv.prepare_features(
             images, cores=cores, converter=converter
         )
 
         featurizer_fp_occ_deriv = GMPFeaturizer(
-            GMPs=GMPs,
-            calc_derivatives=True,
-            calc_occ_derivatives=True,
-            verbose=False,
+            GMPs=GMPs, calc_derivatives=True, calc_occ_derivatives=True, verbose=False,
         )
         features_fp_occ_deriv = featurizer_fp_occ_deriv.prepare_features(
             images, cores=cores, converter=converter
@@ -103,10 +94,7 @@ class FeatruizerTest(unittest.TestCase):
     def test_features_ref(self):
 
         GMPs = {
-            "GMPs": {
-                "orders": [-1, 0, 1, 2, 3],
-                "sigmas": [0.2, 0.4, 0.6, 0.8, 1.0],
-            },
+            "GMPs": {"orders": [-1, 0, 1, 2, 3], "sigmas": [0.2, 0.4, 0.6, 0.8, 1.0],},
             "psp_path": "./test_files/NC-SR.gpsp",
             "square": False,
             "solid_harmonics": True,
@@ -119,10 +107,7 @@ class FeatruizerTest(unittest.TestCase):
     def test_features_method_0(self):
 
         GMPs = {
-            "GMPs": {
-                "orders": [-1, 0, 1, 2, 3],
-                "sigmas": [0.2, 0.4, 0.6, 0.8, 1.0],
-            },
+            "GMPs": {"orders": [-1, 0, 1, 2, 3], "sigmas": [0.2, 0.4, 0.6, 0.8, 1.0],},
             "psp_path": "./test_files/NC-SR.gpsp",
             "square": False,
             "solid_harmonics": True,
@@ -135,10 +120,7 @@ class FeatruizerTest(unittest.TestCase):
     def test_features_method_1(self):
 
         GMPs = {
-            "GMPs": {
-                "orders": [-1, 0, 1, 2, 3],
-                "sigmas": [0.2, 0.4, 0.6, 0.8, 1.0],
-            },
+            "GMPs": {"orders": [-1, 0, 1, 2, 3], "sigmas": [0.2, 0.4, 0.6, 0.8, 1.0],},
             "psp_path": "./test_files/NC-SR.gpsp",
             "square": False,
             "solid_harmonics": True,
@@ -151,10 +133,7 @@ class FeatruizerTest(unittest.TestCase):
     def test_features_method_2(self):
 
         GMPs = {
-            "GMPs": {
-                "orders": [-1, 0, 1, 2, 3],
-                "sigmas": [0.2, 0.4, 0.6, 0.8, 1.0],
-            },
[…]
