### 2023-05-20 90b5889
moved scaling parameter function to GMP class
