+++ GMPFeaturizer/GMP/__init__.py
@@ -20,55 +20,6 @@ from ..util import (
 from ._libgmp import ffi, lib
 
 
-def _get_scaling_constant_radial(sigma):
-    return (1.0 / (sigma * np.sqrt(2.0 * np.pi))) ** 3
-
-
-def _get_scaling_constant_both_probes(order, sigma):
-    alpha = 1 / (2 * sigma * sigma)
-    const = math.pow((math.pi / (2 * alpha)), 1.5)
-    if order == 0:
-        temp = const
-        return 1 / math.sqrt(temp)
-    if order == 1:
-        temp = 0.75 * math.pow(alpha, -1) * const
-        return 1 / math.sqrt(temp)
-    if order == 2:
-        temp = 5.625 * math.pow(alpha, -2) * const
-        return 1 / math.sqrt(temp)
-    if order == 3:
-        temp = 147.65625 * math.pow(alpha, -3) * const
-        return 1 / math.sqrt(temp)
-    if order == 4:
-        temp = 9302.34375 * math.pow(alpha, -4) * const
-        return 1 / math.sqrt(temp)
-    if order == 5:
-        temp = 1151165.0390625 * math.pow(alpha, -5) * const
-        return 1 / math.sqrt(temp)
-    if order == 6:
-        temp = 246924900.87890625 * math.pow(alpha, -6) * const
-        return 1 / math.sqrt(temp)
-    if order == 7:
-        temp = 84263122424.92676 * math.pow(alpha, -7) * const
-        return 1 / math.sqrt(temp)
-    if order == 8:
-        temp = 42974192436712.65 * math.pow(alpha, -8) * const
-        return 1 / math.sqrt(temp)
-    if order == 9:
-        temp = 3.163831150894136e16 * math.pow(alpha, -9) * const
-        return 1 / math.sqrt(temp)
-    raise NotImplementedError
-
-
-def get_scaling_constant(order, sigma, mode):
-    if mode == "radial":
-        return _get_scaling_constant_radial(sigma)
-    elif mode == "both":
-        return _get_scaling_constant_both_probes(order, sigma)
-    else:
-        raise NotImplementedError
-
-
 class GMP(BaseFeature):
     """
     Class for GMP feature
@@ -475,6 +426,69 @@ class GMP(BaseFeature):
         return
 
     # *******************************************************
+    # related to scaling parameter calcualtion
+    # *******************************************************
+
+    def _get_scaling_constant_radial(self, sigma):
+        """
+        private method for calculating scaling parameters
+        based on normalizing the radial probe
+        """
+        return (1.0 / (sigma * np.sqrt(2.0 * np.pi))) ** 3
+
+
+    def _get_scaling_constant_both_probes(self, order, sigma):
+        """
+        private method for calculating scaling parameters
+        based on normalizing the whole probe
+        """
+        alpha = 1 / (2 * sigma * sigma)
+        const = math.pow((math.pi / (2 * alpha)), 1.5)
+        if order == 0:
+            temp = const
+            return 1 / math.sqrt(temp)
+        if order == 1:
+            temp = 0.75 * math.pow(alpha, -1) * const
+            return 1 / math.sqrt(temp)
+        if order == 2:
+            temp = 5.625 * math.pow(alpha, -2) * const
+            return 1 / math.sqrt(temp)
+        if order == 3:
+            temp = 147.65625 * math.pow(alpha, -3) * const
+            return 1 / math.sqrt(temp)
+        if order == 4:
+            temp = 9302.34375 * math.pow(alpha, -4) * const
+            return 1 / math.sqrt(temp)
+        if order == 5:
+            temp = 1151165.0390625 * math.pow(alpha, -5) * const
[…]
