### 2023-06-02 #5476 @bdice
I opened an issue on the repo with some review comments: https://github.com/TRI-AMDD/GMP-featurizer/issues/3

I'd like to see some of those addressed before completing the Review Checklist, since I suspect that some of the proposed changes will affect more than one box on the checklist. Thanks for your submission @RayLei-TRI and thanks for editing @rkurchin!
