### 2023-06-05 TRI-AMDD/GMP-featurizer#3 @JosephMontoya-TRI
Thanks for the review @bdice, awesome feedback.

Just wanted to drop a note to indicate I have agreed to be a co-author, and believe I fulfill the "active project direction" aspect of this codebase.
