### 2023-06-05 TRI-AMDD/GMP-featurizer#3 @bdice
@JosephMontoya-TRI Awesome, thanks!
