### 2023-06-05 a80315d
Update gmp-featurizer-test-main.yml
Try 3.9-3.11 for tests.
