+++ .github/workflows/gmp-featurizer-test-main.yml
@@ -13,7 +13,7 @@ jobs:
           macos-latest,
           # windows-latest
         ]
-        python-version: [3.7, 3.8, 3.9]
+        python-version: [3.9, 3.10, 3.11]
 
     runs-on: ${{ matrix.os }}
 
