### 2023-06-05 6c0411e
Update gmp-featurizer-test-main.yml
Fix yaml syntax to use 3.10 instead of 3.1
