### 2023-06-05 d05a3ec
Update gmp-featurizer-test-main.yml
