### 2023-06-05 d2e7ac1
Update gmp-featurizer-lint.yml
Remove 3.7/3.8 from linting, no need for more versions.
