### 2023-06-05 TRI-AMDD/GMP-featurizer#3 @RayLei-TRI
Thanks @bdice  for the detailed comments and suggestions!
I'm working on the tasks, and will update here
