+++ paper.md
@@ -29,9 +29,9 @@ Representing the local and global environments in atomic systems in a descriptiv
 
 # Overview
 
-The GMP-Featurizer package is mainly written in C++ and python. C++ is used for the underlying computation module for speed, and python used for an intuitive and readable API for scientists and researchers in the community to use. Although this package does not explicitly depend on ASE [@ase-paper] and pymatgen [@pymatgen], two of the most widely used python libraries in the chemistry and materials science communities, APIs are provided so atomic systems defined in these libraries can be easily read by gmp-featurizer. On top of that, Ray is used to parallelize the feature computation, and the parallelization efficiency is close to 100\%.  Overall, this package is designed to be lightweight, easy to use, fast and accurate.
+The GMP-Featurizer package is mainly written in C++ and Python. C++ is used for the underlying computation module for speed, and Python used for an intuitive and readable API for scientists and researchers in the community to use. Two of the most widely used Python libraries in chemistry and materials science are ASE [@ase-paper] and pymatgen [@pymatgen]. The gmp-featurizer library provides APIs that can read atomic systems as input from these libraries, though they are not required as dependencies. This allows the users to leverage the I/O functionalities of these two packages to get data into the format that's accepted by GMP-Featurizer. On top of that, Ray is used to parallelize the feature computation, and the parallelization efficiency is close to 100\%.  Overall, this package is designed to be lightweight, easy to use, fast and accurate.
 
-The main inputs of the workflow are a python dictionary that contains the necessary hyperparameters for defining the desired GMP feature set, and a list of atomic systems that needs to be featured. The native way of defining atomic systems is simply a python dictionary that contains information like `lattice vectors`, `atom positions`, `atom types`, etc. As mentioned, the package also supports both ASE Atoms and pymatgen Structure objects with pre-defined converters. This capability is extensible to other formats with custom-made converters. It also supports the featurization of disordered atomic structures, which is unsupported by many popular featurization methods. Please refer to section \ref{sec:Examples} for more details. The output is simply a list of dictionaries containing the resulting features, and their derivatives if requested, for the atomic structures.
+The main inputs of the workflow are a Python dictionary that contains the necessary hyperparameters for defining the desired GMP feature set, and a list of atomic systems that needs to be featured. The native way of defining atomic systems is simply a Python dictionary that contains information like `lattice vectors`, `atom positions`, `atom types`, etc. As mentioned, the package also supports both ASE Atoms and pymatgen Structure objects with pre-defined converters. This capability is extensible to other formats with custom-made converters. It also supports the featurization of disordered atomic structures, which is unsupported by many popular featurization methods. Please refer to the examples for more details. The output is simply a list of dictionaries containing the resulting features, and their derivatives if requested, for the atomic structures.
 
 By default, the package computes GMP features at each atom position, but it can also be used to compute the features at any set of reference points inside the atomic system by providing a list of the positions of interest for each atomic structure. Users can also specify the number of cores for parallel computing. Moreover, computed results can be cached locally for convenient reprocessing of datasets, e.g. after augmentation or modification. Two sets of standard pseudopotential files are also provided, which are necessary to specify GMP feature sets, but may be difficult to collect from either commercial or open-source density functional theory systems. Lastly, a series of tutorials are provided in the repository to help users with quick starting and understanding the various features of the codebase.
 
