+++ paper.bib
@@ -141,11 +141,9 @@
 	author = {Collins, C. and Dyer, M. S. and Pitcher, M. J. and Whitehead, G. F. S. and Zanella, M. and Mandal, P. and Claridge, J. B. and Darling, G. R. and Rosseinsky, M. J.},
 	month = jun,
 	year = {2017},
-	note = {Number: 7657
-Publisher: Nature Publishing Group},
+	note = {Number: 7657, Publisher: Nature Publishing Group},
 	keywords = {Materials chemistry, Theoretical chemistry, Theory and computation},
 	pages = {280--284},
-	file = {Full Text PDF:/Users/josephmontoya/Zotero/storage/XWUZ839H/Collins et al. - 2017 - Accelerated discovery of two crystal structure typ.pdf:application/pdf},
 }
 
 @misc{amptorch,
@@ -167,4 +165,19 @@ Publisher: Nature Publishing Group},
     publisher = {American Physical Society},
     doi = {10.1103/PhysRevLett.98.146401},
     url = {https://link.aps.org/doi/10.1103/PhysRevLett.98.146401}
+}
+
+@inproceedings{Ray,
+    author = {Moritz, Philipp and Nishihara, Robert and Wang, Stephanie and Tumanov, Alexey and Liaw, Richard and Liang, Eric and Elibol, Melih and Yang, Zongheng and Paul, William and Jordan, Michael I. and Stoica, Ion},
+    title = {Ray: A Distributed Framework for Emerging AI Applications},
+    year = {2018},
+    isbn = {9781931971478},
+    publisher = {USENIX Association},
+    address = {USA},
+    abstract = {The next generation of AI applications will continuously interact with the environment and learn from these interactions. These applications impose new and demanding systems requirements, both in terms of performance and flexibility. In this paper, we consider these requirements and present Ray--a distributed system to address them. Ray implements a unified interface that can express both task-parallel and actor-based computations, supported by a single dynamic execution engine. To meet the performance requirements, Ray employs a distributed scheduler and a distributed and fault-tolerant store to manage the system's control state. In our experiments, we demonstrate scaling beyond 1.8 million tasks per second and better performance than existing specialized systems for several challenging reinforcement learning applications.},
+    booktitle = {Proceedings of the 13th USENIX Conference on Operating Systems Design and Implementation},
+    pages = {561–577},
+    numpages = {17},
+    location = {Carlsbad, CA, USA},
+    series = {OSDI'18}
 }
