+++ CONTRIBUTING.md
@@ -1,0 +1,61 @@
+# Welcome to GMP-Featurizer contributing guide <!-- omit in toc -->
+
+Thank you for interesting in investing your time and contributing to our project! 
+
+In this guide you will get an overview of the contribution workflow from opening an issue, creating a PR, reviewing, and merging the PR.
+
+
+## New contributor guide
+
+To get an overview of the project, read the [README](README.md) and the associated [short paper](paper.md).
+
+Please refer to the examples for the functionalities of this
+
+
+
+### Issues
+
+#### Create a new issue
+
+If you spot a problem with the package, please look through if it's mentioned in any of the issues. If a related issue doesn't exist, you can open a new issue in the repository.
+
+#### Feature request
+
+You are welcomed to submit feature requests by opening up new issues. We will try our best to meet everybody's need. In the issue, please describe your request clearly, and tell us why you think such feature would be a good addition to the package
+
+#### Solve an issue
+
+Scan through our current issues to find one that interests you. We don’t assign issues to anyone, so if you find an issue to work on, you are welcome to open a PR with a fix.
+
+
+### Make Changes
+
+
+1. Fork the repository.
+- Using GitHub Desktop:
+  - [Getting started with GitHub Desktop](https://docs.github.com/en/desktop/installing-and-configuring-github-desktop/getting-started-with-github-desktop) will guide you through setting up Desktop.
+  - Once Desktop is set up, you can use it to [fork the repo](https://docs.github.com/en/desktop/contributing-and-collaborating-using-github-desktop/cloning-and-forking-repositories-from-github-desktop)!
+
+- Using the command line:
+  - [Fork the repo](https://docs.github.com/en/github/getting-started-with-github/fork-a-repo#fork-an-example-repository) so that you can make your changes without affecting the original project until you're ready to merge them.
+
+2. Create a working branch and start with your changes.
+
+### Commit your update
+
+Commit the changes once you are happy with them. Don't forget to self-review to speed up the review process:zap:.
+
+### Pull Request
+
+When you're finished with the changes, create a pull request, also known as a PR.
+- Fill the "Ready for review" template so that we can review your PR. This template helps reviewers understand your changes as well as the purpose of your pull request.
+- Don't forget to [link PR to issue](https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue) if you are solving one.
+- Enable the checkbox to [allow maintainer edits](https://docs.github.com/en/github/collaborating-with-issues-and-pull-requests/allowing-changes-to-a-pull-request-branch-created-from-a-fork) so the branch can be updated for a merge.
+Once you submit your PR, a Docs team member will review your proposal. We may ask questions or request additional information.
+- We may ask for changes to be made before a PR can be merged, either using [suggested changes](https://docs.github.com/en/github/collaborating-with-issues-and-pull-requests/incorporating-feedback-in-your-pull-request) or pull request comments. You can apply suggested changes directly through the UI. You can make any other changes in your fork, then commit them to your branch.
+- As you update your PR and apply changes, mark each conversation as [resolved](https://docs.github.com/en/github/collaborating-with-issues-and-pull-requests/commenting-on-a-pull-request#resolving-conversations).
+- If you run into any merge issues, checkout this [git tutorial](https://github.com/skills/resolve-merge-conflicts) to help you resolve merge conflicts and other issues.
+
+### Your PR is merged!
+
+Congratulations :tada::tada: The GMP-Featurizer team thanks you :sparkles:.
