### 2023-06-07 TRI-AMDD/GMP-featurizer#3 @RayLei-TRI
@bdice Thanks again for the great suggestions and comments!
We have so far addressed most of them (except the documentation and testing workflow, which we are working on now), please see the updated master branch. We will also keep you updated here.

One great point you mentioned is about the examples and what downstream tasks are possible.  We have added a simple example for training a simple ML model to the atomic Mulliken charge with GMP features. This is of course not the only application, some important applications include using it to train machine learning force field (as described in the original GMP paper), or train to other atom-resolved or spatially-resolved properties. However, these researches are still ongoing so we think it's outside the scope of this package for now, but will add them in when appropriate. On the other hand, we give the users the option to cache the computed features in case they need to access them afterward.  For example when the features need to be accessed again in downstream/different applications, or when there is a system crash and features calculated and stored in memory could be lost. It is of course not always needed, just an option.
