+++ README.md
@@ -66,7 +66,7 @@ GMPs = {
     # "square": False, 
 }
 
-featurizer = GMPFeaturizer(GMPs=GMPs, calc_derivatives=True, verbose=True)
+featurizer = GMPFeaturizer(GMPs=GMPs, converter=converter, calc_derivatives=True, verbose=True)
 ```
 Set calc_derivatives=True if you want to get the feature derivatives w.r.t. atom positions, which are stored in the form of sparse matrices.
 
@@ -75,7 +75,7 @@ Set calc_derivatives=True if you want to get the feature derivatives w.r.t. atom
 #### Calculate features and access data
 Use the "cores" argument to change the number of cores for parallelization. Also converted needed to be specified,
 ```
-result = featurizer.prepare_features(images, cores=5, converter=converter)
+result = featurizer.prepare_features(images, cores=5)
 
 features = [entry["features"] for entry in result]
 feature_primes = [entry["feature_primes"] for entry in result]
@@ -119,12 +119,12 @@ GMPs = {
     #no need to change if you are not considering the feature derivatives
     # "square": False, 
 }
-featurizer = GMPFeaturizer(GMPs=GMPs, calc_derivatives=True, verbose=True)
+featurizer = GMPFeaturizer(GMPs=GMPs, converter=converter, calc_derivatives=True, verbose=True)
 
 
 
 # calculate features
-result = featurizer.prepare_features(images, cores=5, converter=converter)
+result = featurizer.prepare_features(images, cores=5)
 
 # access data
 features = [entry["features"] for entry in result]
@@ -136,8 +136,8 @@ Simply set "save_features=True" when calling the prepare_features function.
 
 The path to the local database is set when initializing the featurizer
 ```
-featurizer = GMPFeaturizer(GMPs=GMPs, calc_derivatives=False, feature_database="cache/features/")
-features = featurizer.prepare_features(images, cores=5, save_features=True, converter=converter)
+featurizer = GMPFeaturizer(GMPs=GMPs, converter=converter, calc_derivatives=False, feature_database="cache/features/")
+features = featurizer.prepare_features(images, cores=5, save_features=True)
 ```
 
 #### License
