+++ tests/featurizer_test.py
@@ -17,33 +17,34 @@ class FeatruizerTest(unittest.TestCase):
 
         featurizer_noderiv = GMPFeaturizer(
             GMPs=GMPs,
+            converter=converter,
             calc_derivatives=False,
             calc_occ_derivatives=False,
             verbose=False,
         )
         features_noderiv = featurizer_noderiv.prepare_features(
-            images, cores=cores, converter=converter
+            images, cores=cores
         )
 
         featurizer_fp_deriv = GMPFeaturizer(
             GMPs=GMPs, calc_derivatives=True, calc_occ_derivatives=False, verbose=False,
         )
         features_fp_deriv = featurizer_fp_deriv.prepare_features(
-            images, cores=cores, converter=converter
+            images, cores=cores
         )
 
         featurizer_occ_deriv = GMPFeaturizer(
             GMPs=GMPs, calc_derivatives=False, calc_occ_derivatives=True, verbose=False,
         )
         features_occ_deriv = featurizer_occ_deriv.prepare_features(
-            images, cores=cores, converter=converter
+            images, cores=cores
         )
 
         featurizer_fp_occ_deriv = GMPFeaturizer(
             GMPs=GMPs, calc_derivatives=True, calc_occ_derivatives=True, verbose=False,
         )
         features_fp_occ_deriv = featurizer_fp_occ_deriv.prepare_features(
-            images, cores=cores, converter=converter
+            images, cores=cores
         )
 
         assert np.all(
