+++ docs/source/GMPFeaturizer.rst
@@ -1,0 +1,30 @@
+GMPFeaturizer package
+=====================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 7
+
+   GMPFeaturizer.GMP
+
+Submodules
+----------
+
+.. toctree::
+   :maxdepth: 7
+
+   GMPFeaturizer.base_feature
+   GMPFeaturizer.constants
+   GMPFeaturizer.converters
+   GMPFeaturizer.gmp_featurizer
+   GMPFeaturizer.util
+
+Module contents
+---------------
+
+.. automodule:: GMPFeaturizer
+   :members:
+   :undoc-members:
+   :show-inheritance:
