+++ docs/source/index.rst
@@ -1,0 +1,20 @@
+.. GMP-featurizer documentation master file, created by
+   sphinx-quickstart on Thu Jun  8 17:05:01 2023.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to GMP-featurizer's documentation!
+==========================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
