### 2023-06-09 78e22ff
update docs for proper copyright and add .nojekyll
