+++ docs_build/build/html/_sources/GMPFeaturizer.gmp_featurizer.rst.txt
@@ -1,7 +1,0 @@
-GMPFeaturizer.gmp\_featurizer module
-====================================
-
-.. automodule:: GMPFeaturizer.gmp_featurizer
-   :members:
-   :undoc-members:
-   :show-inheritance:
