+++ docs_build/build/html/_sources/GMPFeaturizer.GMP.rst.txt
@@ -1,18 +1,0 @@
-GMPFeaturizer.GMP package
-=========================
-
-Submodules
-----------
-
-.. toctree::
-   :maxdepth: 7
-
-   GMPFeaturizer.GMP.libgmp_builder
-
-Module contents
----------------
-
-.. automodule:: GMPFeaturizer.GMP
-   :members:
-   :undoc-members:
-   :show-inheritance:
