+++ docs_build/build/html/_sources/GMPFeaturizer.base_feature.rst.txt
@@ -1,7 +1,0 @@
-GMPFeaturizer.base\_feature module
-==================================
-
-.. automodule:: GMPFeaturizer.base_feature
-   :members:
-   :undoc-members:
-   :show-inheritance:
