+++ docs/html/_sources/index.rst.txt
@@ -1,5 +1,5 @@
 .. GMP-featurizer documentation master file, created by
-   sphinx-quickstart on Thu Jun  8 17:05:01 2023.
+   sphinx-quickstart on Thu Jun  8 17:55:40 2023.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
