### 2023-06-12 #5476 @isdanni
>

thanks for checking in! author(@RayLei-TRI) has addressed my previous issues with paper/documentation. I'm confirming the final package performance and function tests.
