+++ docs_build/source/conf.py
@@ -13,7 +13,9 @@ author = 'Toyota Research Institute'
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-extensions = []
+extensions = [
+    "sphinx.ext.napoleon"
+]
 
 templates_path = ['_templates']
 exclude_patterns = []
