### 2023-06-14 TRI-AMDD/GMP-featurizer#3 @RayLei-TRI
@bdice Thanks again for the suggestions!
I think we have completed all the listed tasks, and the repo is definitely better now.
Please let us know what you think, or if you have any other concerns.
