### 2023-06-14 TRI-AMDD/GMP-featurizer#3 @bdice
Thank you! I will be out for a few days but I can review this again next week. I’ll set a reminder. 👍
