### 2023-06-21 TRI-AMDD/GMP-featurizer#8 @yw-fang
[REVIEW comments] for the paper and code submitted to JOSS
Hello, @RayLei-TRI and @JosephMontoya-TRI  sorry for my late review report of the manuscript submitted to Journal of Open Source Software https://github.com/openjournals/joss-reviews/issues/5476

Here are my minor comments which you may consider during revision

- [x] I personally think the requirements.txt is not complete. To test the code, I built a new virtual environment and installed the code following the instructions. When testing the examples, the code reminded me that I needed to install ase, numpy, and so on. I realized some of these libraries were included in the requirements.txt of earlier versions, could authors explain why did they remove it? To make the code ready completely, I would suggest including all the necessary and major libraries in the requirements.txt

- [x] In the manuscript, when referring to the "cffi" library, it may be more appropriate to use "Cffi" or "CFFI" to distinguish it as a specific terminology. Upon reading the summary at the beginning, I mistakenly assumed it was a typographical error.

- [x] The reference of ray in the manuscript was not formatted properly (e.g. "Ray?" appears in the current summary).

- [x] I noticed that CIF files taken from ICSD couldn't been read by ase_read. However, if I convert the CIF from ICSD to a CIF from VESTA software. It works. I know this issue doesn't from GMP-featurizer itself, but you may consider raising a warning if read of CIF fails. Note that I only test one CIF from ICSD, and you may cross-test it for confirmation.

- [x] In the section Overview of manuscript, "Please refer to section ??" should be fixed.

- [x] Authors are invited to create a webpage to document the tutorials and manuals.
