### 2023-06-22 TRI-AMDD/GMP-featurizer#8 @RayLei-TRI
Hi @yw-fang , Thank you very much for your great and detailed suggestions! We have updated the package accordingly, and here are our responses:

1) We took out ASE and pymatgen from the requirement list, because they are not required to run the featurizer itself, and adding them as requirements might make the package cumbersome. However, since they are popular packages for the likely users of GMP-Featurizer, we provided converters to interface GMP-featurizer with them (e.g. for IO purposes).  However, it is a little strange that numpy was not installed since it was in the requirement.txt file, and I was not able to reproduce the problem. Would you mind sharing more detail about the problem? It would be great if you could include information about the OS, python version, error details, and we will try to reproduce and fix the bug.

2) Thanks for the suggestion! we have updated it to CFFI.

3) Thanks for catching it!  We fixed it earlier on the main branch but did not merge the change into the JOSS branch until now. Now it should be fixed.

4) Thanks for the suggestion! Would you mind sharing more information about the error? Did ase_read() fail silently, and GMP-featurizer errored out because of it? It would be great if you could share with us the example file you used, and we will try to reproduce the error and update the package accordingly.

5) Same as 3), it should be fixed now, thanks a lot for catching it!

6) Thanks for the suggestion! We have created a minimal working website for hosting the docs and manuals: https://tri-amdd.github.io/GMP-featurizer/html/index.html

Thanks again for your review comments!
