### 2023-06-22 TRI-AMDD/GMP-featurizer#8 @yw-fang
Hello @RayLei-TRI thanks for your quick response.
My comments are as follows

> Hi @yw-fang , Thank you very much for your great and detailed suggestions! We have updated the package accordingly, and here are our responses:
>
> 1. We took out ASE and pymatgen from the requirement list, because they are not required to run the featurizer itself, and adding them as requirements might make the package cumbersome. However, since they are popular packages for the likely users of GMP-Featurizer, we provided converters to interface GMP-featurizer with them (e.g. for IO purposes).  However, it is a little strange that numpy was not installed since it was in the requirement.txt file, and I was not able to reproduce the problem. Would you mind sharing more detail about the problem? It would be great if you could include information about the OS, python version, error details, and we will try to reproduce and fix the bug.

It makes sense, since you'd like to make the code lightweight by excluding the heavy libraries. As for numpy, it's my fault. The command 'pip install requirements.txt' can install numpy as expected.

> 2. Thanks for the suggestion! we have updated it to CFFI.
> 3. Thanks for catching it!  We fixed it earlier on the main branch but did not merge the change into the JOSS branch until now. Now it should be fixed.
> 4. Thanks for the suggestion! Would you mind sharing more information about the error? Did ase_read() fail silently, and GMP-featurizer errored out because of it? It would be great if you could share with us the example file you used, and we will try to reproduce the error and update the package accordingly.

Sure. Here is the error information:

```
--------------------------------------------------------------------------
RuntimeError                              Traceback (most recent call last)
Cell In[3], line 6
      3 from ase.io import read as aseread
      5 # Loading cif file as a ase atoms object
----> 6 image = aseread("./SmFeO3.cif")
      7 # The input to the featurizer should be a non-empty list
      8 images = [image]

File ~/miniconda3/envs/py38/lib/python3.8/site-packages/ase/io/formats.py:736, in read(filename, index, format, parallel, do_not_split_by_at_sign, **kwargs)
    733     return list(_iread(filename, index, format, io, parallel=parallel,
    734                        **kwargs))
    735 else:
--> 736     return next(_iread(filename, slice(index, None), format, io,
    737                        parallel=parallel, **kwargs))

File ~/miniconda3/envs/py38/lib/python3.8/site-packages/ase/parallel.py:275, in parallel_generator.<locals>.new_generator(*args, **kwargs)
    269 @functools.wraps(generator)
    270 def new_generator(*args, **kwargs):
    271     if (world.size == 1 or
    272         args and getattr(args[0], 'serial', False) or
    273         not kwargs.pop('parallel', True)):
    274         # Disable:
--> 275         for result in generator(*args, **kwargs):
    276             yield result
    277         return

File ~/miniconda3/envs/py38/lib/python3.8/site-packages/ase/io/formats.py:803, in _iread(filename, index, format, io, parallel, full_output, **kwargs)
    801 # Make sure fd is closed in case loop doesn't finish:
    802 try:
--> 803     for dct in io.read(fd, *args, **kwargs):
    804         if not isinstance(dct, dict):
    805             dct = {'atoms': dct}

File ~/miniconda3/envs/py38/lib/python3.8/site-packages/ase/io/cif.py:604, in read_cif(fileobj, index, store_tags, primitive_cell, subtrans_included, fractional_occupancies, reader)
    602 # Find all CIF blocks with valid crystal data
    603 images = []
--> 604 for block in parse_cif(fileobj, reader):
    605     if not block.has_structure():
    606         continue

File ~/miniconda3/envs/py38/lib/python3.8/site-packages/ase/io/cif.py:537, in parse_cif_ase(fileobj)
    534 if not line or line.startswith('#'):
    535     continue
--> 537 yield parse_block(lines, line)

File ~/miniconda3/envs/py38/lib/python3.8/site-packages/ase/io/cif.py:500, in parse_block(lines, line)
    498 assert line.lower().startswith('data_')
    499 blockname = line.split('_', 1)[1].rstrip()
--> 500 tags = parse_items(lines, line)
    501 return CIFBlock(blockname, tags)

File ~/miniconda3/envs/py38/lib/python3.8/site-packages/ase/io/cif.py:189, in parse_items(lines, line)
    187     tags[key.lower()] = value
    188 elif lowerline.startswith('loop_'):
--> 189     tags.update(parse_loop(lines))
    190 elif lowerline.startswith('data_'):
    191     if line:

File ~/miniconda3/envs/py38/lib/python3.8/site-packages/ase/io/cif.py:161, in parse_loop(lines)
    158 headers = list(parse_cif_loop_headers(lines))
    159 # Dict would be better.  But there can be repeated headers.
--> 161 columns = parse_cif_loop_data(lines, len(headers))
    163 columns_dict = {}
    164 for i, header in enumerate(headers):

File ~/miniconda3/envs/py38/lib/python3.8/site-packages/ase/io/cif.py:149, in parse_cif_loop_data(lines, ncolumns)
    147 if tokens:
    148     assert len(tokens) < ncolumns
--> 149     raise RuntimeError('CIF loop ended unexpectedly with incomplete row')
    151 return columns

RuntimeError: CIF loop ended unexpectedly with incomplete row
```

You may reproduce the error using this CIF from ICSD database:  https://ehubox.ehu.eus/s/QnrJX25JGA435Gt  (the link expires on June 24 Spain time)

> 5. Same as 3), it should be fixed now, thanks a lot for catching it!
> 6. Thanks for the suggestion! We have created a minimal working website for hosting the docs and manuals: https://tri-amdd.github.io/GMP-featurizer/html/index.html

It appears to be a promising start! I'll keep an eye on it! Thanks for your contribution to the community.

>
> Thanks again for your review comments!
