### 2023-06-29 #5476 @bdice
How do I get a rendered PDF of the latest paper commit?

I looked at https://github.com/openjournals/joss-papers/blob/joss.05476/joss.05476/10.21105.joss.05476.pdf which is linked above (https://github.com/openjournals/joss-reviews/issues/5476#issuecomment-1552975799) but it seems to be out of sync with this source: https://github.com/TRI-AMDD/GMP-featurizer/blob/JOSS/paper.md
