### 2023-06-29 TRI-AMDD/GMP-featurizer#3 @bdice
Thanks for your many improvements to this repository! This is much better than last time I reviewed.

I have a few remaining comments:

1. The README mentions a file `test.cif` in its example snippet, but it's not clear where this file is located or how to acquire it. The README example snippet uses `ase` but doesn't instruct the user to install it, nor is it a dependency of this package.

2. I built the documentation locally. It's rather sparse. Perhaps you can help guide users to the main classes in the API docs by writing some content on the index page. Also I see this warning when building:

```
GMP-featurizer/docs/source/modules.rst: WARNING: document isn't included in any toctree
```

3. I'm not sure how to get the latest paper to render. I commented here about that: https://github.com/openjournals/joss-reviews/issues/5476#issuecomment-1613871776
