### 2023-06-29 #5476 @bdice
I have checked most of the boxes in my review and left a comment summarizing my comments since last time. https://github.com/TRI-AMDD/GMP-featurizer/issues/3#issuecomment-1613872044

I'd like to see a revised PDF of the paper before approving my review.
