### 2023-07-12 TRI-AMDD/GMP-featurizer#8 @RayLei-TRI
Hi @yw-fang, Thank you for the comment and sorry for the late reply.
It seems the IO error was cost by the ase io module, and it printed out the details of the error as well, so I think it might not make sense to address it in this package. However, I did add a sample cif file in the examples folder so that users don't have to find their own to just try the package, hope this address your concern
Thanks again, and please let me know if you have any additional comments :)
