### 2023-07-12 TRI-AMDD/GMP-featurizer#8 @yw-fang
> Hi @yw-fang, Thank you for the comment and sorry for the late reply. It seems the IO error was cost by the ase io module, and it printed out the details of the error as well, so I think it might not make sense to address it in this package. However, I did add a sample cif file in the examples folder so that users don't have to find their own to just try the package, hope this address your concern Thanks again, and please let me know if you have any additional comments :)

I agree. As I mentioned initially, this issue doesn't come from your code itself. I didn't expect you to solve it here. But for convenience, you can include a warning or something like it in README or any other proper document. I'm fine with your response.
I'll tick all the tasks, and recommend the present version to be published.
Of course, feel free to close this Github issue.

Just to correlate https://github.com/openjournals/joss-reviews/issues/5476 I have completed the review and recommend it for publication.
