### 2023-07-17 #5476 @bdice
@RayLei-TRI Looks like some of the boxes were checked for fixes I requested, but the requested changes aren't showing up in the latest draft PDF above. Let me know if you disagree with my suggestions, otherwise I'll wait to approve until those changes are reflected. Thanks for your work on this, hopefully the remaining pieces are quick to address. 👍
