+++ docs_build/source/conf.py
@@ -14,7 +14,8 @@ author = 'Toyota Research Institute'
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
 extensions = [
-    "sphinx.ext.napoleon"
+    "sphinx.ext.napoleon",
+    "sphinx.ext.autosectionlabel"
 ]
 
 templates_path = ['_templates']
@@ -27,3 +28,10 @@ exclude_patterns = []
 
 html_theme = 'alabaster'
 html_static_path = ['_static']
+pygments_style = 'sphinx'
+html_sidebars = {
+    '**': [
+        'localtoc.html',
+        'searchbox.html',
+    ]
+}
