### 2023-07-18 TRI-AMDD/GMP-featurizer#3 @RayLei-TRI
Hi @bdice, hope you had a great weekend and sorry about the belated response
we have just updated the code accordingly. an example cif file is added, and we have updated the webpage with some more content. The compiled can be found [here](https://github.com/openjournals/joss-reviews/issues/5476#issuecomment-1615277983)

Thanks again for your comments, and please let us know if you have any further comments!
