+++ paper.md
@@ -25,7 +25,7 @@ GMP-Featurizer is a lightweight, accurate, efficient, and scalable software pack
 
 # Statement of need
 
-Representing the local and global environments in atomic systems in a descriptive and efficient way has been an important research top in the chemistry, chemical engineering, and material science communities. Having good representations, or features, of chemical environments has proven to be vital for building reliable machine learning (ML) models. These models can accurately predict properties of atomic systems, and in limited cases have even been used for discovering or designing new chemicals and materials [@zuo_accelerating_2021; @collins_accelerated_2017]. So far, scientists and researchers have designed featurization schemes like the atom-centered symmetry function (ACSF) [@BehlerParrinello], the smooth overlap of atomic positions (SOAP) [@SOAP] and the Gaussian Momentum [@Gaussian_Momentum] schemes. More recently, graph representation and ML models based on them (e.g. MEGNet [@MEGNet], CGCNN [@CGCNN]) have been successful. Gaussian Multipole, or GMP [@GMP], is a recently developed scheme of featurizing local chemical environments, i.e. the chemical characteristics of spaces near individual atoms in molecules and crystal structures. GMP approximates underlying local electronic environments (e.g. approximated distribution of local electron cloud) using multipole expansion, the theory of which is explicated in a prior publication [@GMP]. The featurization scheme is flexible, depending only on prior assumptions of atomic identity and position, and it is therefore applicable to various atomic systems (molecules, nano-particles, periodic crystals, etc.) in which atomic arrangements are known. Feature computation is fast, and the representation accuracy is systematically improvable. Moreover, thanks to the deep connection between the GMP features and physics, we have previously shown that ML models based on these features are transferable [@GMP]. With these characteristics, GMP featurization could be useful to a broad audience for future research in chemistry and materials science. Therefore, having lightweight, reliable, and open software that can calculate these features in a fast and accurate way is desirable.
+Representing the local and global environments in atomic systems in a descriptive and efficient way has been an important research top in the chemistry, chemical engineering, and material science communities. Having good representations, or features, of chemical environments has proven to be vital for building reliable machine learning (ML) models. These models can accurately predict properties of atomic systems, and in limited cases have even been used for discovering or designing new chemicals and materials [@zuo_accelerating_2021] [@collins_accelerated_2017]. So far, scientists and researchers have designed featurization schemes like the atom-centered symmetry function (ACSF) [@BehlerParrinello], the smooth overlap of atomic positions (SOAP) [@SOAP] and the Gaussian Momentum [@Gaussian_Momentum] schemes. More recently, graph representation and ML models based on them (e.g. MEGNet [@MEGNet], CGCNN [@CGCNN]) have been successful. Gaussian Multipole, or GMP [@GMP], is a recently developed scheme of featurizing local chemical environments, i.e. the chemical characteristics of spaces near individual atoms in molecules and crystal structures. GMP approximates underlying local electronic environments (e.g. approximated distribution of local electron cloud) using multipole expansion, the theory of which is explicated in a prior publication [@GMP]. The featurization scheme is flexible, depending only on prior assumptions of atomic identity and position, and it is therefore applicable to various atomic systems (molecules, nano-particles, periodic crystals, etc.) in which atomic arrangements are known. Feature computation is fast, and the representation accuracy is systematically improvable. Moreover, thanks to the deep connection between the GMP features and physics, we have previously shown that ML models based on these features are transferable [@GMP]. With these characteristics, GMP featurization could be useful to a broad audience for future research in chemistry and materials science. Therefore, having lightweight, reliable, and open software that can calculate these features in a fast and accurate way is desirable.
 
 # Overview
 
