### 2023-07-18 TRI-AMDD/GMP-featurizer#3 @bdice
I opened https://github.com/TRI-AMDD/GMP-featurizer/pull/10 with a couple small suggestions, then I'll approve my review. Thanks for the changes you've implemented!
