### 2023-07-18 #5476 @rkurchin
Editorial comments:
- line 27: merge references into one so there aren't two separate sets of parentheses
- 37: replace "explicated" with "introduced" or similar
- 39: nanoparticle doesn't need a hyphen
- 54: rephrase to " [...] into the format required by [...]"
- 57: missing comma after "fast"
- 60: "featured" should be "featurized"
- 71: be consistent about capitalization - either GMP-featurizer or GMP-Featurizer
- 75: "autonomously" should be "automatically"
- 78: move "greatly" to after "vary"
- 79-80: don't capitalize chemical element names
