+++ paper.bib
@@ -16,7 +16,7 @@
     keywords = {SciAdv r-articles ; Physical Sciences ; Physics},
 }
 
-@article{Gaussian_Momentum,
+@article{GaussianMomentum,
     author = {Zaverkin, V. and Kästner, J.},
     title = {Gaussian Moments as Physically Inspired Molecular Descriptors for Accurate and Scalable Machine Learning Potentials},
     journal = {Journal of Chemical Theory and Computation},
@@ -108,7 +108,7 @@
     abstract={The atomic simulation environment (ASE) is a software package written in the Python programming language with the aim of setting up, steering, and analyzing atomistic simulations. In ASE, tasks are fully scripted in Python. The powerful syntax of Python combined with the NumPy array library make it possible to perform very complex simulation tasks. For example, a sequence of calculations may be performed with the use of a simple ‘for-loop’ construction. Calculations of energy, forces, stresses and other quantities are performed through interfaces to many external electronic structure codes or force fields using a uniform interface. On top of this calculator interface, ASE provides modules for performing many standard simulation tasks such as structure optimization, molecular dynamics, handling of constraints and performing nudged elastic band calculations.}
 }
 
-@article{zuo_accelerating_2021,
+@article{ZuoAccelerated2021,
 	title = {Accelerating materials discovery with {Bayesian} optimization and graph deep learning},
 	volume = {51},
 	issn = {1369-7021},
@@ -126,7 +126,7 @@
 	file = {ScienceDirect Full Text PDF:/Users/josephmontoya/Zotero/storage/RD2HZ9WN/Zuo et al. - 2021 - Accelerating materials discovery with Bayesian opt.pdf:application/pdf;ScienceDirect Snapshot:/Users/josephmontoya/Zotero/storage/FDUV895X/S1369702121002984.html:text/html},
 }
 
-@article{collins_accelerated_2017,
+@article{CollinsAccelerated2017,
 	title = {Accelerated discovery of two crystal structure types in a complex inorganic phase field},
 	volume = {546},
 	copyright = {2017 Macmillan Publishers Limited, part of Springer Nature. All rights reserved.},
