### 2023-07-19 #5476 @RayLei-TRI
@rkurchin I have made the edits as you suggested, cleared all the typos, double checked the authors, affiliations and license, as well as made the release on github. The only thing left to do is archiving it using Zenodo, which I do need organizational approval, since the repo belongs to our organization. Hopefully this won't take too long, and I will keep you updated.

Please let me know if you have any other concerns.
