### 2023-07-21 #5476 @RayLei-TRI
@rkurchin I have archived it on Zenodo: 10.5281/zenodo.8170029
the author names and ORCIDs are double-checked.
The only question I have is the license: the license of the package is Apache 2.0, but it's not an option on Zenodo, it only has "Creative Commons Attribution ..." licenses. Please advice. Thank you so much!
