### 2023-07-21 #5476 @rkurchin
> The only question I have is the license: the license of the package is Apache 2.0, but it's not an option on Zenodo, it only has "Creative Commons Attribution ..." licenses. Please advice. Thank you so much!

@openjournals/bcm-eics, thoughts on @RayLei-TRI's question here?
