### 2023-07-25 #5476 @rkurchin
Pinging @openjournals/bcm-eics again about the license question above, this submission is otherwise ready to be accepted as far as I'm concerned!
