+++ docs_build/source/conf.py
@@ -6,32 +6,23 @@
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = 'GMP-featurizer'
-copyright = '2023, Toyota Research Institute'
-author = 'Toyota Research Institute'
+project = "GMP-featurizer"
+copyright = "2023, Toyota Research Institute"
+author = "Toyota Research Institute"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-extensions = [
-    "sphinx.ext.napoleon",
-    "sphinx.ext.autosectionlabel"
-]
+extensions = ["sphinx.ext.napoleon", "sphinx.ext.autosectionlabel"]
 
-templates_path = ['_templates']
+templates_path = ["_templates"]
 exclude_patterns = []
-
 
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
-html_theme = 'alabaster'
-html_static_path = ['_static']
-pygments_style = 'sphinx'
-html_sidebars = {
-    '**': [
-        'localtoc.html',
-        'searchbox.html',
-    ]
-}
+html_theme = "alabaster"
+html_static_path = ["_static"]
+pygments_style = "sphinx"
+html_sidebars = {"**": ["localtoc.html", "searchbox.html",]}
