### 2023-07-29 41cb909
major update, correct bugs in cutoff calcualtion and spatial partitioning
