+++ GMPFeaturizer/GMP/__init__.py
@@ -10,7 +10,7 @@ import numpy as np
 from scipy import sparse
 
 from ..base_feature import BaseFeature
-from ..constants import ATOM_SYMBOL_TO_INDEX_DICT
+from ..constants import ATOM_SYMBOL_TO_INDEX_DICT, ATOM_LIST
 from ..util import (
     _gen_2Darray_for_ffi,
     list_symbols_to_indices,
@@ -228,10 +228,15 @@ class GMP(BaseFeature):
 
             if order == -1:
                 element_distances = []
-                for element, psp in self.atomic_psp.items():
-                    element_distance = self._get_default_cutoff_single_element(
-                        0, psp, threshold=self.overlap_threshold
-                    )
+                # for element, psp in self.atomic_psp.items():
+                for element in ATOM_LIST:
+                    try:
+                        psp = self.atomic_psp[element]
+                        element_distance = self._get_default_cutoff_single_element(
+                            0, psp, threshold=self.overlap_threshold
+                        )
+                    except:
+                        element_distance = 0
                     element_distances.append(element_distance)
                 elemental_order_sigma_cutoffs.append(element_distances)
                 order_sigma_index_dict[(-1, 0)] = order_sigma_index
@@ -239,10 +244,15 @@ class GMP(BaseFeature):
 
             else:
                 element_distances = []
-                for element, psp in self.atomic_psp.items():
-                    element_distance = self._get_default_cutoff_single_element(
-                        sigma, psp, threshold=self.overlap_threshold * factors[order],
-                    )
+                # for element, psp in self.atomic_psp.items():
+                for element in ATOM_LIST:
+                    try:
+                        psp = self.atomic_psp[element]
+                        element_distance = self._get_default_cutoff_single_element(
+                            sigma, psp, threshold=self.overlap_threshold * factors[order],
+                        )
+                    except:
+                        element_distance = 0
                     element_distances.append(element_distance)
                     # print(order, sigma, element, element_distance)
                 elemental_order_sigma_cutoffs.append(element_distances)
@@ -302,10 +312,15 @@ class GMP(BaseFeature):
             if order == -1:
                 element_gaussian_distances = []
                 element_distances = []
-                for element, psp in self.atomic_psp.items():
-                    gaussian_distances = self._get_default_cutoff_element_gaussians(
-                        0, psp, self.max_num_gaussians, threshold=self.overlap_threshold
-                    )
+                # for element, psp in self.atomic_psp.items():
+                for element in ATOM_LIST:
+                    try: 
+                        psp = self.atomic_psp[element]
+                        gaussian_distances = self._get_default_cutoff_element_gaussians(
+                            0, psp, self.max_num_gaussians, threshold=self.overlap_threshold
+                        )
+                    except:
+                        gaussian_distances = [0] * self.max_num_gaussians
                     element_distance = np.max(gaussian_distances).item()
                     element_distances.append(element_distance)
                     element_gaussian_distances += gaussian_distances
@@ -320,13 +335,19 @@ class GMP(BaseFeature):
             else:
                 element_gaussian_distances = []
                 element_distances = []
-                for element, psp in self.atomic_psp.items():
-                    gaussian_distances = self._get_default_cutoff_element_gaussians(
-                        sigma,
-                        psp,
-                        self.max_num_gaussians,
-                        threshold=self.overlap_threshold * factors[order],
-                    )
+                for element in ATOM_LIST:
+                    try: 
+                        psp = self.atomic_psp[element]
[…]
