+++ GMPFeaturizer/GMP/__init__.py
@@ -249,7 +249,9 @@ class GMP(BaseFeature):
                     try:
                         psp = self.atomic_psp[element]
                         element_distance = self._get_default_cutoff_single_element(
-                            sigma, psp, threshold=self.overlap_threshold * factors[order],
+                            sigma,
+                            psp,
+                            threshold=self.overlap_threshold * factors[order],
                         )
                     except:
                         element_distance = 0
@@ -314,10 +316,13 @@ class GMP(BaseFeature):
                 element_distances = []
                 # for element, psp in self.atomic_psp.items():
                 for element in ATOM_LIST:
-                    try: 
+                    try:
                         psp = self.atomic_psp[element]
                         gaussian_distances = self._get_default_cutoff_element_gaussians(
-                            0, psp, self.max_num_gaussians, threshold=self.overlap_threshold
+                            0,
+                            psp,
+                            self.max_num_gaussians,
+                            threshold=self.overlap_threshold,
                         )
                     except:
                         gaussian_distances = [0] * self.max_num_gaussians
@@ -336,10 +341,10 @@ class GMP(BaseFeature):
                 element_gaussian_distances = []
                 element_distances = []
                 for element in ATOM_LIST:
-                    try: 
+                    try:
                         psp = self.atomic_psp[element]
-                # for element, psp in self.atomic_psp.items():
-                    # print(element)
+                        # for element, psp in self.atomic_psp.items():
+                        # print(element)
                         gaussian_distances = self._get_default_cutoff_element_gaussians(
                             sigma,
                             psp,
@@ -371,7 +376,6 @@ class GMP(BaseFeature):
 
         # print("Shape of elemental_order_sigma_gaussian_cutoffs: ", elemental_order_sigma_gaussian_cutoffs.shape)
         # print("Shape of elemental_order_sigma_cutoffs: ", elemental_order_sigma_cutoffs.shape)
-
 
         self.params_set[
             "elemental_order_sigma_gaussian_cutoffs"
@@ -412,7 +416,7 @@ class GMP(BaseFeature):
             element_distances = []
             # for element, psp in self.atomic_psp.items():
             for element in ATOM_LIST:
-                try: 
+                try:
                     psp = self.atomic_psp[element]
                 except:
                     element_distance = self._get_default_cutoff_single_element(
@@ -611,20 +615,24 @@ class GMP(BaseFeature):
                 len(self.atomic_gaussian_setup[element_index]) / 2
             )
             # ngaussian_list.append(self.params_set[element_index]["gaussian_count"])
-            ngaussian_list[element_index] = self.params_set[element_index]["gaussian_count"]
+            ngaussian_list[element_index] = self.params_set[element_index][
+                "gaussian_count"
+            ]
             # print("self.params_set[element_index]: {}".format(self.params_set[element_index]))
 
         # ngaussian_list = np.asarray(ngaussian_list, dtype=np.intc, order="C")
         max_gaussian_count = np.max(ngaussian_list)
-        overall_gaussian_params = np.zeros((120, max_gaussian_count * 2), dtype=np.float64, order="C")
+        overall_gaussian_params = np.zeros(
+            (120, max_gaussian_count * 2), dtype=np.float64, order="C"
+        )
         for element_index in self.element_indices:
             # temp = np.zeros(max_gaussian_count * 2)
             # temp[
             #     : self.params_set[element_index]["gaussian_count"] * 2
             # ] = self.params_set[element_index]["gaussian_params"]
             # overall_gaussian_params.append(temp)
-            overall_gaussian_params[element_index, 
-                : self.params_set[element_index]["gaussian_count"] * 2
+            overall_gaussian_params[
[…]
