### 2023-08-01 5cb2226
change try-except to if-else
