### 2023-08-09 #5476 @Kevin-Mattheus-Moerman
@RayLei-TRI I have checked this myself (also needed to edit a license for one of my own projects). Please check the following:

1. Make sure that `Upload type` is set to `Software`
![Screenshot from 2023-08-09 13-37-22](https://github.com/openjournals/joss-reviews/assets/8392709/588f92db-69a7-4ba3-b306-4a2ff350d286)

2. Next, in the `License` field you can start typing the name of your license, i.e. `Apache 2.0`, and select it.
![Screenshot from 2023-08-09 13-35-57](https://github.com/openjournals/joss-reviews/assets/8392709/5cf79aa1-1168-403f-b118-f338337de761)

3. Next save these changes, and that should do it.

Let me know if you have any more difficulties.
